Rename laika package 'laika'
